feat(frontend): created notifications box, removed the original page
